can see all categories and can see profile page - refactored profile and user view
