+++ pages/users/edit/[id].js
@@ -3,19 +3,19 @@ import { useEffect, useState } from 'react';
 import { getSingleUser } from '../../../utils/data/userData';
 import RegisterForm from '../../../components/RegisterForm';
 
-export default function EditProfile() {
-  const [editProfile, setEditProfile] = useState({});
+export default function EditUser() {
+  const [editUser, setEditUser] = useState({});
   const router = useRouter();
   const { id } = router.query;
 
   useEffect(() => {
-    getSingleUser(id).then(setEditProfile);
+    getSingleUser(id).then(setEditUser);
   }, [id]);
 
   return (
     <div>
       <h1>Edit Profile</h1>
-      <RegisterForm user={editProfile} />
+      <RegisterForm user={editUser} />
     </div>
   );
 }
